requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow-cpu==2.19.0
numpy==2.0.2
Pillow==11.3.0
gunicorn==22.0.0
